+++ libs/core/src/main/java/org/opensearch/core/util/BytesRefUtils.java
@@ -102,4 +102,13 @@ public class BytesRefUtils {
         }
         return uniqueCount;
     }
+
+    public static long bytesToLong(BytesRef bytes) {
+        int high = (bytes.bytes[bytes.offset + 0] << 24) | ((bytes.bytes[bytes.offset + 1] & 0xff) << 16) | ((bytes.bytes[bytes.offset + 2]
+            & 0xff) << 8) | (bytes.bytes[bytes.offset + 3] & 0xff);
+        int low = (bytes.bytes[bytes.offset + 4] << 24) | ((bytes.bytes[bytes.offset + 5] & 0xff) << 16) | ((bytes.bytes[bytes.offset + 6]
+            & 0xff) << 8) | (bytes.bytes[bytes.offset + 7] & 0xff);
+        return (((long) high) << 32) | (low & 0x0ffffffffL);
+    }
+
 }